refactor(formatting): extract BSL formatting engine from FormatProvider
Движок форматирования (getNewText + хелперы, ~460 строк) жил в providers.FormatProvider,
хотя это переиспользуемая логика, а не тонкий LSP-провайдер. Из-за прямого вызова движка
диагностикой IdenticalExpressionsDiagnostic (нормализация текста операнда для сообщения)
возникала инверсия diagnostics→providers.

Движок вынесен в новый бин formatting.BslFormatter; FormatProvider стал тонким
LSP-адаптером и делегирует в него (как требует providers/CLAUDE.md). Диагностика теперь
зависит от BslFormatter, а не от providers. Поведение не изменилось — FormatProviderTest
(27 тестов) и IdenticalExpressionsDiagnosticTest.checkMessage() зелёные.

Это убрало последнюю слоистую инверсию, поэтому правило layer_dependencies снято с
FreezingArchRule и стало строгим (без базовой линии). В archunit_store осталась только
база по @Autowired.
