--- archunit_store/84350e2f-67b7-4a66-b776-0b0d67f02341
@@ -1,3 +1,0 @@
-Constructor <com.github._1c_syntax.bsl.languageserver.diagnostics.IdenticalExpressionsDiagnostic.<init>(com.github._1c_syntax.bsl.languageserver.providers.FormatProvider)> has parameter of type <com.github._1c_syntax.bsl.languageserver.providers.FormatProvider> in (IdenticalExpressionsDiagnostic.java:0)
-Field <com.github._1c_syntax.bsl.languageserver.diagnostics.IdenticalExpressionsDiagnostic.formatProvider> has type <com.github._1c_syntax.bsl.languageserver.providers.FormatProvider> in (IdenticalExpressionsDiagnostic.java:0)
-Method <com.github._1c_syntax.bsl.languageserver.diagnostics.IdenticalExpressionsDiagnostic.getOperandText(com.github._1c_syntax.bsl.languageserver.utils.expressiontree.BinaryOperationNode)> calls method <com.github._1c_syntax.bsl.languageserver.providers.FormatProvider.getNewText(java.util.List, java.util.Locale, org.eclipse.lsp4j.Range, int, org.eclipse.lsp4j.FormattingOptions)> in (IdenticalExpressionsDiagnostic.java:197)
