+++ archunit_store/stored.rules
@@ -1,4 +1,3 @@
 #
-#Fri Jun 26 16:05:45 UTC 2026
-\u041F\u043E\u043B\u044F\ \u0438\ \u043C\u0435\u0442\u043E\u0434\u044B\ \u043D\u0435\ \u0434\u043E\u043B\u0436\u043D\u044B\ \u043F\u043E\u043C\u0435\u0447\u0430\u0442\u044C\u0441\u044F\ @Autowired\ (\u043F\u0440\u0435\u0434\u043F\u043E\u0447\u0442\u0438\u0442\u0435\u043B\u044C\u043D\u043E\ \u043A\u043E\u043D\u0441\u0442\u0440\u0443\u043A\u0442\u043E\u0440\u043D\u043E\u0435\ \u0432\u043D\u0435\u0434\u0440\u0435\u043D\u0438\u0435)=0bcc5045-6c2a-4645-9d43-af332e79e4a2
-\u0421\u043B\u043E\u0438\u0441\u0442\u0430\u044F\ \u0430\u0440\u0445\u0438\u0442\u0435\u043A\u0442\u0443\u0440\u0430\:\ cli\ \u0438\ providers\ \u043D\u0435\ \u0438\u0441\u043F\u043E\u043B\u044C\u0437\u0443\u044E\u0442\u0441\u044F\ \u00AB\u0441\u043D\u0438\u0437\u0443\u00BB\ (\u044F\u0434\u0440\u043E\u043C/diagnostics)=84350e2f-67b7-4a66-b776-0b0d67f02341
+#Fri Jun 26 16:52:41 UTC 2026
+\u041F\u043E\u043B\u044F\ \u0438\ \u043C\u0435\u0442\u043E\u0434\u044B\ \u043D\u0435\ \u0434\u043E\u043B\u0436\u043D\u044B\ \u043F\u043E\u043C\u0435\u0447\u0430\u0442\u044C\u0441\u044F\ @Autowired\ (\u043F\u0440\u0435\u0434\u043F\u043E\u0447\u0442\u0438\u0442\u0435\u043B\u044C\u043D\u043E\ \u043A\u043E\u043D\u0441\u0442\u0440\u0443\u043A\u0442\u043E\u0440\u043D\u043E\u0435\ \u0432\u043D\u0435\u0434\u0440\u0435\u043D\u0438\u0435)=1c379082-d0df-444c-91e9-3226009069dd
